+++ CHANGELOG.rst
@@ -35,6 +35,7 @@ Changed
 - Added single backquotes default to code mode in docs (:gh:`379` by `Julian Tachella`_)
 - Changed the __add__ method for stack method for stacking physics (:gh:`371` by `Julian Tachella`_ and `Andrew Wang`_)
 - Changed the R2R loss to handle multiple noise distributions (:gh:`380` by `Brayan Monroy`_)
+- `Trainer.get_samples_online` using physics generator now updates physics params via both `update_parameters` and forward pass (:gh:`386` by `Andrew Wang`_)
 - Deprecate Trainer freq_plot in favour of plot_interval (:gh:`388` by `Andrew Wang`_)
 
 v0.2.2